pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastweigh-ops-intelligence"
version = "0.1.0"
description = "Operations Intelligence Pack for Fast-Weigh GraphQL V2"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "FastWeigh Intelligence" }]
dependencies = [
  "apscheduler>=3.10.4",
  "duckdb>=1.1.3",
  "graphql-core>=3.2.6",
  "httpx>=0.27.2",
  "pandas>=2.2.3",
  "plotly>=5.24.1",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.5.2",
  "python-dateutil>=2.9.0.post0",
  "PyYAML>=6.0.2",
  "streamlit>=1.39.0",
  "tenacity>=9.0.0",
  "typer>=0.12.5"
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11.2",
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.9"
]

[project.scripts]
ops-intel = "ops_intelligence.cli:app"

[tool.setuptools]
packages = ["ops_intelligence", "ops_intelligence.graphql", "ops_intelligence.extraction", "ops_intelligence.warehouse", "ops_intelligence.dashboard", "ops_intelligence.alerts", "ops_intelligence.reports", "ops_intelligence.scheduler"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
ignore_missing_imports = true
